requirements.txt
Flask>=3.0.0
opencv-python>=4.8.0
face-recognition>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
SQLAlchemy>=2.0.35
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0
PyMySQL>=1.1.0
cryptography>=3.0.0